Switch as2gbmap over to Python 3
We don't ship Python 2 any more, so either we need to drop this or
convert it. Doesn't look like there's anything particularly Python 2
specific, so clean up some basic formatting errors and changed the
shebang to python3. Still drop it from /usr/bin however, so we don't
have to depend on Python3 for a single script.